% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{adap.surf}
\alias{adap.surf}
\title{Calculate combined surface calculate W on the adaptive landscape
 W = w1*F1 + w2*F2 (Polly et al, 2016)}
\usage{
adap.surf(Fn, wn, xmar, ymar, n, ...)
}
\arguments{
\item{Fn}{A multi.surf object from multi.fnc.surface}

\item{wn}{A set of performance weights. Can be user derived or generated
using search.w.exhaustive}

\item{xmar, ymar}{A vector of length 2 defining the X and Y limits}

\item{n}{The number of grid points to subset along each axis}

\item{...}{Parameters to pass onto nested functions}
}
\value{
A polynomial function and surface defining the summed adaptive
  landscape
}
\description{
Calculate combined surface calculate W on the adaptive landscape
 W = w1*F1 + w2*F2 (Polly et al, 2016)
}
