% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{anc_scores}
\alias{anc_scores}
\title{Compute node pc scores using ancestral state reconstruction}
\usage{
anc_scores(scores, tree, drop = F)
}
\arguments{
\item{scores}{pc scores for tip taxa}

\item{tree}{a tree object}

\item{drop}{logical. automatically drop missing tip taxa from tree. defaults to FALSE}
}
\value{
returns node pc scores
}
\description{
Compute node pc scores using ancestral state reconstruction
}
\examples{
X
}
