% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{calc.best.wn}
\alias{calc.best.wn}
\title{Calculate the best model fits}
\usage{
calc.best.wn(X, sortby = "lik", percentile = 0.99, method = c("quantile"))
}
\arguments{
\item{X}{A weights matrix generated from search.w.exhaustive()}

\item{sortby}{c("z", "lik", "dis"). Which optimization vector to sort by.
Defaults to "lik"}

\item{percentile}{percentile (alpha) cuttoff. Defaults to 0.99 (0.01)}

\item{method}{method by which to calculate top fits. Method = "quantile"
simply caclulates the average weights from top percentile set by the user.
Method = "chi-squared" calculates the best significant models based on the
log-likelihood ratio distribution.}
}
\value{
A list containing wn the best averaged model weights, wn.se the
  standard error, wn.sd te standard deviation and wn.range, the range of top
  wn values
}
\description{
Calculate the best model fits
}
\examples{
X
}
\author{
Blake Dickson and Katrina Jones
}
