% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lscp_Par.r
\name{calc_Ri}
\alias{calc_Ri}
\title{Function for calculating Pareto optimality, using the formula of Deakin et al}
\usage{
calc_Ri(x, n1, n2)
}
\arguments{
\item{x}{The grid data from objects of class 'kriged_surface'
or 'by_Wprime'}

\item{n1}{Name of the first trait/group to be compared}

\item{n2}{Name of the second trait/group to be compared}
}
\value{
Pareto optimality
}
\description{
Function for calculating Pareto optimality, using the formula of Deakin et al
}
\examples{
Standard' optimality, get the rankings of each point on the grid
Inverse' optimality, get the inverse rankings of each point on the grid
}
