% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{fnc_dataframe}
\alias{fnc_dataframe}
\title{Coerces XY coordinate data and covariate data into a list, or optionally into
   a 3D array}
\usage{
fnc_dataframe(x, y, z, row.names, func.names = NULL, array = F, scale = T)
}
\arguments{
\item{x, y}{A numeric vectors with x,y coordinates of observations in
morphospace}

\item{z}{A Numeric vector or matrix containing covariate observations corresponding to
x,y coordinates. Can contain more than one covariate}

\item{func.names}{An optional string containing the names for traits given
in z}

\item{array}{An optional logical argument to return a 3D array rather than a
list. By default is set to FALSE}

\item{scale}{An optional logical argument to normalize z covariates to the
same unit scale. Defaults to TRUE. Can be set to FALSE if the user wishes
to normalize manually}
}
\value{
Returns a list, or optionally an Array. Each sheet contains an x,y,z
    functional dataframe corresponding to covariates given in z,
    labelled by 'func.names'
}
\description{
Coerces XY coordinate data and covariate data into a list, or optionally into
   a 3D array
}
