% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{generate.weights}
\alias{generate.weights}
\title{Generate a generic dataframe containing weight combinations. Can be used in
  search.w.exhaustive, or to split jobs for distributed computing}
\usage{
generate.weights(step, nvar, varnames = NULL, time.est = F, verbose = T, ...)
}
\arguments{
\item{step}{Numeric. Degree by which to vary weights by.}

\item{nvar}{Numeric. Number of variables (columns) to generate}

\item{varnames}{Optional. Names for variables}

\item{time.est}{Optional. Estimate length of time to perform
search.w.exhaustive}

\item{verbose}{Logical. Return dataframe. Used for silencing in time.est}

\item{...}{}
}
\value{

}
\description{
Generate a generic dataframe containing weight combinations. Can be used in
  search.w.exhaustive, or to split jobs for distributed computing
}
\examples{
X
}
