% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.r
\name{get_grid}
\alias{get_grid}
\title{get_grid}
\usage{
get_grid(x)
}
\arguments{
\item{x}{A "by_Wprime" or "kriged_surfaces" object}
}
\value{
The grid of x, y values and other data, with the
		addition of a unique grid cell identifier, 'gridID', which
		is needed to correctly reorder the data following the ranking
		functions - see below.
}
\description{
Given a "by_Wprime" or "kriged_surfaces" object, this
function returns the grid of x, y values and adds a
unique grid cell identifier
}
