% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{init.QTM.cluster}
\alias{init.QTM.cluster}
\title{Initialize a distributed analysis of trait weights}
\usage{
init.QTM.cluster(
  Fn.surfaces,
  morphospace,
  dir,
  weights = NULL,
  step = NULL,
  njobs
)
}
\arguments{
\item{Fn.surfaces}{A dataframe containing trait performance surfaces
generated from multi.fnc.surface}

\item{morphospace}{A pca object defining the morphospace}

\item{dir}{Directory to store cluster analysis objects}

\item{weights}{Optional. A dataframe containing weights from generate.weights}

\item{step}{Optional. If weights is not provided, specify step size}

\item{njobs}{Number of jobs to split into}
}
\value{
A directory of split jobs
}
\description{
Initialize a distributed analysis of trait weights
}
\examples{
X
}
