% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lands.grp.test}
\alias{lands.grp.test}
\title{Significance test between group weights}
\usage{
lands.grp.test(grpa, grpb, percentile = 0.01)
}
\arguments{
\item{grpa, grpb}{Weights dataframes generated from search.w.exhaustive()}

\item{percentile}{Numeric. Determines the top percentile of group weights. Defaults to 0.01}
}
\value{

}
\description{
Significance test between group weights
}
\examples{
X
}
\author{
Katrina Jones
}
