% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{lscp.pairwise}
\alias{lscp.pairwise}
\title{Pairwise significance testing between group model fits}
\usage{
lscp.pairwise(X1, X2, p.val = 0.99)
}
\arguments{
\item{X1, X2}{Dataframe containing Wn model fits for two groups}

\item{p.val}{}
}
\value{
List containing p.values and matching weights
}
\description{
Pairwise significance testing between group model fits
}
\examples{
X
}
\author{
Katrina Jones
}
