% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{multi.fnc.surface}
\alias{multi.fnc.surface}
\title{Produces a list of polynomial surface fits from multiple functional
    dataframes. Is essentailly an apply wrapper for fnc.surface}
\usage{
multi.fnc.surface(X, npoly = 3, npoints = NULL)
}
\arguments{
\item{X}{A list of x,y,z functonal dataframes from 'fnc_dataframe'}

\item{npoly}{Numeric. Determines the degree of polynomial to apply to the
surface. Usually 3rd order polynomials produce the best fit}

\item{npoints}{Numeric. Optional argument to up or downsample the number x,y
grid points. Defaults to the number of observations in X}
}
\value{
Returns a multi.fnc.surface object. A list containing N number of
    functional surfaces. See fnc.surface for details
}
\description{
Produces a list of polynomial surface fits from multiple functional
    dataframes. Is essentailly an apply wrapper for fnc.surface
}
