% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{multi.lands.grp.test}
\alias{multi.lands.grp.test}
\title{Tests for pairwise signifince between mulitple group weights}
\usage{
multi.lands.grp.test(groups, percentile = 0.01)
}
\arguments{
\item{groups}{A list of named group weights from search.w.exhaustive()}

\item{percentile}{Numeric. Determines the top percentile of group weights. Defaults to 0.01}
}
\value{

}
\description{
Tests for pairwise signifince between mulitple group weights
}
\examples{
X
}
\author{
Katrina Jones
}
