% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{pareto}
\alias{pareto}
\title{Calculate pareto front between two landscapes}
\usage{
pareto(surf.1, surf.2, plot = F, a = NULL, smooth = c("kernel"), ...)
}
\arguments{
\item{surf.1, surf.2}{surface objects}

\item{plot}{logical, plot pareto front}

\item{a}{sequence from 1-0 to calculate the pareto transtion}

\item{smooth}{optional c("spline", "kernel"), determines which smoothing
method to use. The "spline" method will fail if mulitple values of Y exist
for a given X, in which case use "kernel"}

\item{...}{additional parameters for plot.surf}
}
\value{
a list containing the pareto transition
}
\description{
Calculate pareto front between two landscapes
}
\examples{
X
}
