% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscapes.R
\name{plot.image}
\alias{plot.image}
\title{Plot a heatmap of functional data.}
\usage{
\method{plot}{image}(x, y, z, ...)
}
\arguments{
\item{x, y, z}{x can be an xyz dataframe (such as from fnc.df),
or the user can provide separate xyz vectors}

\item{...}{parameters to pass onto image()}
}
\description{
Plot a heatmap of functional data.
}
\examples{
X
}
