% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscapes.R
\name{plot.surf}
\alias{plot.surf}
\title{General plotting function to plot surfaces. Can be used to plot fnc.surfaces,
   adapt.surfaces or just a morpohpscae}
\usage{
\method{plot}{surf}(
  surf,
  method = "poly",
  contour = T,
  contour.lines = NULL,
  nlevels = 40,
  points = NULL,
  tree = NULL,
  axes = T,
  node.points = NULL,
  max.point = T,
  palette = viridis,
  box = F,
  pch = 21,
  pt.col = NULL,
  ...
)
}
\arguments{
\item{surf}{A $surface object from fnc.surface or adap.surface.}

\item{method}{"can be "poly" or "kriging" Plot surface as a least-squares polynomial fit, or as a kriging surface}

\item{contour}{logical. Plot coloured contours onto morphospace.
Defaults to TRUE}

\item{contour.lines}{Optional. Colour to set contour lines. Defaults to NULL.}

\item{nlevels}{Number of colour levels to plot contours}

\item{points}{Optional. Provide points to plot on top of surface}

\item{tree}{Optional. Plot a tree on top of surface}

\item{axes}{Logical. Plot axis ticks. Defaults to TRUE}

\item{node.points}{points for plotting node. Required if tree is provided}

\item{max.point}{Optional logical. Plot highest point on surface. Defaults to
TRUE}

\item{palette}{Optional colour palette. Defaults to viridis colour palette.}

\item{box}{Logical. Plot box around surface}

\item{pch}{set point character. Defaults to 21}

\item{pt.col}{set point colour.}

\item{...}{other parameters to pass onto plot()}
}
\description{
General plotting function to plot surfaces. Can be used to plot fnc.surfaces,
   adapt.surfaces or just a morpohpscae
}
\examples{
X
}
