% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lscp_Par.r
\name{plot_ggsurf}
\alias{plot_ggsurf}
\title{plot_ggsurf}
\usage{
plot_ggsurf(x, n1 = NULL)
}
\arguments{
\item{x}{Dataframe or tibble containing trait values}

\item{v1}{Name of the trait variable or group landscape being plotted}
}
\value{
A ggplot object
}
\description{
Helper function - plots specific performance surface or landscape. Data
must be in tabular format. Grid is color-coded according to trait value
or Z score.
}
