% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscapes.R
\name{plot_multisurf}
\alias{plot_multisurf}
\title{Plot multiple surfaces}
\usage{
plot_multisurf(multi.surf, main = NULL, ...)
}
\arguments{
\item{multi.surf}{A multi-surface pobject from multi.fnc.surface, or a list of adaptive surfaces}

\item{...}{optional parameters to pass onti plot_surf and par}
}
\value{

}
\description{
Plot multiple surfaces
}
\examples{
X
}
