% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscapes.R
\name{plotKrige}
\alias{plotKrige}
\title{Title}
\usage{
plotKrige(X, points = NA, ...)
}
\arguments{
\item{X}{a Kriged surface object from krige_surf}

\item{...}{options to pass to spplot}

\item{col.pal}{}
}
\value{

}
\description{
Title
}
