% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prob_peak}
\alias{prob_peak}
\title{Probability of belonging to a groups adaptive peak}
\usage{
prob_peak(surf, train, test, n = 1000, plot = F)
}
\arguments{
\item{surf}{a surface object from and adaptive landscape}

\item{train}{pc scores from test data. Usually the same data used to optimize the landscape}

\item{test}{pc scores for data to test}

\item{n}{number of rnorm samples}

\item{plot}{plot z-height distributions for training and test data.}
}
\value{
a list with z-height probabilities and the 95% z-height CI for the landscape
}
\description{
Probability of belonging to a groups adaptive peak
}
