% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{reduce_ans}
\alias{reduce_ans}
\title{recombines completed split jobs}
\usage{
reduce_ans(ans.dir, sortby = c("z", "lik", "dist"))
}
\arguments{
\item{ans.dir}{Directory containing compoleted jobs}
}
\value{
returns a combined matrix of all computed weight combinations
}
\description{
recombines completed split jobs
}
\examples{
X
}
