% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{reduce_dir}
\alias{reduce_dir}
\title{Title}
\usage{
reduce_dir(ans.dir, result.dir, sortby = c("z", "lik", "dist"))
}
\arguments{
\item{ans.dir}{}

\item{result.dir}{}

\item{sortby}{}
}
\value{

}
\description{
Title
}
\examples{
X
}
