% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.r
\name{sp_vals_from_grid}
\alias{sp_vals_from_grid}
\title{sp_vals_from_grid}
\usage{
sp_vals_from_grid(x, y)
}
\arguments{
\item{x}{A 'by_Wprime' or 'PO_List' object - contains
the functional adaptive landscapes or Pareto landscapes
you want to extract heights from.}

\item{y}{A dataframe containing species information morphospace. The coordinates
columns must be labelled 'x' and 'y'.}
}
\value{
A dataframe containing the species information from y,
with additional columns extracted from the landscape data
}
\description{
Function to get the heights on a landscape for a set
of species points.
For a given set of species coordinates in morphospace,
this function finds the closest point on the landscape
grid, and assigns the values associated to that grid
point to that species.
}
