% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{split.jobs}
\alias{split.jobs}
\title{Splits a weights dataframe into a number of jobs to be computed separately}
\usage{
\method{split}{jobs}(weights, njobs, dir = NULL)

\method{split}{jobs}(weights, njobs, dir = NULL)
}
\arguments{
\item{weights}{}

\item{njobs}{}

\item{dir}{}

\item{weights.df}{A dataframe of weight combinations to split.
Can be generated from generate_weights}

\item{out.dir}{The directory to save jobs to}

\item{overwrite}{A logical scalar. Should the output folder be overwitten?
Defaults to FALSE}
}
\value{
Writes njobs number of job files to be computed separatley


}
\description{
Splits a weights dataframe into a number of jobs to be computed separately

Title
}
\examples{
X X
X
}
