% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{split_jobs}
\alias{split_jobs}
\title{Splits a weights dataframe into a number of jobs to be computed separately}
\usage{
split_jobs(weights.df, njobs, out.dir = NULL, overwrite = F)
}
\arguments{
\item{weights.df}{A dataframe of weight combinations to split}

\item{njobs}{The number of jobs to split into}

\item{out.dir}{The directory to save jobs to}

\item{overwrite}{A logical scalar. Should the output folder be overwitten?
Defaults to FALSE}
}
\value{
Writes njobs number of job files to be computed separatley
}
\description{
Splits a weights dataframe into a number of jobs to be computed separately
}
