% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{sum_surface}
\alias{sum_surface}
\title{Sum any number of adaptive landscapes}
\usage{
sum_surface(landscapes)
}
\arguments{
\item{landscapes}{a list of adaptive lanscapes}
}
\value{
Returns a combined adpative landscape
}
\description{
Sum any number of adaptive landscapes
}
\examples{
X
}
