% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{surf_hull}
\alias{surf_hull}
\title{Create alpha hull around data}
\usage{
surf_hull(coords2D, alpha = 0.001, plot = F)
}
\arguments{
\item{coords2D}{a matrix of 2D XY morphospace coordinates}

\item{alpha}{alpha value for creating hulls}

\item{plot}{logical. Plot alpha hull on morphospace}
}
\value{
returns a grid of points within the alpha hull.
}
\description{
Create alpha hull around data
}
