% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{surfaceAlongTree}
\alias{surfaceAlongTree}
\title{Plot a performance surface or adaptive landscape onto a tree}
\usage{
surfaceAlongTree(
  tree,
  tip.scores,
  node.scores = NULL,
  landscape,
  palette.cols = NULL,
  plot = T,
  palette = NULL,
  binary = F,
  ...
)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{tip.scores}{a matrix containing pc.scores for each tipt taxa}

\item{node.scores}{a matrix containing pc. scores for each node}

\item{landscape}{a landscape object from fnc.surface or adap.surface}

\item{palette.cols}{a color palette}

\item{plot}{Logical. Plot trait mapped tree. Defaults to FALSE}

\item{palette}{a palette function}

\item{binary}{Logical. plot transition as a binary or continuous trait.
Defaults to FALSE}

\item{...}{Optional parameters to pass to plot.contMap}
}
\value{
Returns a list contianing tip and node scores, predicted heights on
 the landscape and a cont.mapped tree object
}
\description{
Plot a performance surface or adaptive landscape onto a tree
}
\examples{
X
}
