% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{time.est}
\alias{time.est}
\title{Estimate length of time to perform exhaustive search}
\usage{
\method{time}{est}(num.perm, nvar, Fn, xmar, ymar)
}
\arguments{
\item{num.perm}{number of permutations (rows of weights)}

\item{nvar}{number of variables}

\item{Fn}{multi.func surfaces object}

\item{xmar, ymar}{set X and Y margins}
}
\value{
Returns an estimate time
}
\description{
Estimate length of time to perform exhaustive search
}
\examples{
X
}
