% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{trans.surface}
\alias{trans.surface}
\title{Calculate a transition landscape between two landscapes}
\usage{
trans.surface(X, Y, binary = F)
}
\arguments{
\item{X, Y}{A $surface onject}

\item{binary}{Logical. Calcuate as a binary transition
(STILL IN DEVELOPMENT). Defaults to FALSE}
}
\value{
Returns a $surface object
}
\description{
Calculate a transition landscape between two landscapes
}
\examples{
X
}
