\name{turtles_tree}
\alias{turtles_tree}
\docType{data}
\title{
Trutle Phylogeny
}
\description{
A phylogeny of turtles from Dickson and Pierce (2019) downloaded from timetree.org
}
\usage{data("turtles_tree")}
\format{
  A phylo object that stores the turtle phylogeny.
  \describe{
     }
  }

\source{
Dickson, B.V. and Pierce, S.E. (2019), Functional performance of turtle humerus shape across an ecological adaptive landscape. Evolution, 73: 1265-1277. \doi{10.1111/evo.13747}

timetree.org
}
\keyword{datasets}