% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{fnc.surface}
\alias{fnc.surface}
\title{Produces a 3D polynomial surface of a single functional dataframe.
    Useful for testing and checking data}
\usage{
fnc.surface(
  X,
  method = "poly",
  npoints = NULL,
  plot = F,
  pad = 1.2,
  fnc.name = NULL,
  range = NULL,
  ...
)
}
\arguments{
\item{X}{An x,y,z funtional dataframe produced by fnc_dataframe}

\item{npoints}{Numeric. Optional argument to up or downsample the number x,y
grid points. Defaults to the number of observations in X}

\item{plot}{Logical. Plot surfaces. Defaults to true}

\item{pad}{Add padding to range data. Defaults to adding 20% on all range margins}

\item{fnc.name}{Optional string to label the functional surface}

\item{...}{optional paramaters to pass onto plot}

\item{npoly}{Numeric. Determines the degree of polynomial to apply to the
surface. Usually 3rd order polynomials produce the best fit}
}
\value{
Returns a list of polynomial surface objects:
    poly: the polynomial model applied to X
    summary: the polynomial model summary
    surface: the polynomial surface fit to X
}
\description{
Produces a 3D polynomial surface of a single functional dataframe.
    Useful for testing and checking data
}
\examples{
X
}
