% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{init.QTM.cluster}
\alias{init.QTM.cluster}
\title{Title}
\usage{
init.QTM.cluster(
  Fn.surfaces,
  morphospace,
  dir,
  perms = NULL,
  step = NULL,
  njobs
)
}
\arguments{
\item{Fn.surfaces}{}

\item{morphospace}{}

\item{dir}{}

\item{perms}{}

\item{step}{}

\item{njobs}{}
}
\value{

}
\description{
Title
}
