% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{krige_surf}
\alias{krige_surf}
\title{Surface Kriging}
\usage{
krige_surf(X, hull = T, alpha = 0.001, hullPlot = F)
}
\arguments{
\item{X}{A Functional Dataframe. Columns 1,2 should contain XY morphospace coordinates, with subsequent columns containing functional data correspoinding to morphospace points.}

\item{hull}{optional logical. If TRUE, function will only Krige area within the alpha hull as calculated by surf_hull}

\item{alpha}{optional. Alpha value for surf_hull}

\item{hullPlot}{Logical. Plot hull?}
}
\value{
returns
}
\description{
Surface Kriging
}
