% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{multi.fnc.surface}
\alias{multi.fnc.surface}
\title{Produces a list of polynomial surface fits from multiple functional
    dataframes. Is essentailly an apply wrapper for fnc.surface}
\usage{
multi.fnc.surface(X, par = par(), ...)
}
\arguments{
\item{X}{A data table of XY coordinates and Z trait values, or a list of x,y,z functonal dataframes from 'fnc.dataframe'}

\item{...}{Paramaters to pass onto fnc.surface and plot}
}
\value{
Returns a multi.fnc.surface object. A list containing N number of
    functional surfaces. See fnc.surface for details
}
\description{
Produces a list of polynomial surface fits from multiple functional
    dataframes. Is essentailly an apply wrapper for fnc.surface
}
\examples{
X
}
