% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lscp_Par.r
\name{plot_ggtrade}
\alias{plot_ggtrade}
\title{plot_ggtrade}
\usage{
plot_ggtrade(x, n1 = NULL, n2 = NULL)
}
\arguments{
\item{x}{PO_List object, result of calc_lscp_Pareto. Alternatively, a dataframe or tibble containing trait values}

\item{n1}{First variable to plot}

\item{n2}{Second variable to plot}
}
\value{
A ggplot object
}
\description{
Helper function - plots trade-offs between different traits or
landscape Z scores, as an x-y scatterplot. Points are color coded
according to Pareto optimality Ri score.
}
