% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lscp_Par.r
\name{plot_lscp_Pareto}
\alias{plot_lscp_Pareto}
\title{Utility functions for calculating Pareto optimality}
\usage{
plot_lscp_Pareto(x, n1 = NULL, n2 = NULL)
}
\arguments{
\item{x}{Output from calc_lscps_Pareto}
}
\value{
A series of ggplot graphs, generated by a series of helper functions
}
\description{
Plots Pareto optimality between two traits or landscapes Z values across morphospace.
}
