% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{search.w.exhaustive}
\alias{search.w.exhaustive}
\title{Calculate the optimum weights for each trait given a position, or group in
morphospace}
\usage{
search.w.exhaustive(
  Zprime,
  Fn,
  weights.df = NULL,
  step = NULL,
  xmar = xmar,
  ymar = ymar,
  Cluster = F,
  method = c("mean", "sum"),
  optimum = c("absolute")
)
}
\arguments{
\item{Zprime}{Numeric vector of length=2, or Nx2 matrix containing x,y
coordiates to optimize for}

\item{Fn}{A multi.fnc.surface list containing traits to model for}

\item{weights.df}{An optional matrix containing all possible combinations of
wn given n number of traits.}

\item{step}{Optional numeric. The degree by which to vary wn by. Required if
weights.df is not provided. Smaller step sizes require exponentially more
 computation time.}

\item{xmar, ymar}{Numeric vector containing min/max margins for morphospace}

\item{Cluster}{logical. Split computation across multiple machines.
Currently not working}

\item{method}{Method by which to summarize optimization results when multiple
Zprime coordintes are provided. Both produce similar results, but have
 not been stress tested.}

\item{optimum}{Criteria to optimize for and sort results by. Currently only
"absolute" height on the landscape is supported, though all optimization
metrics are returned. See details.}
}
\value{
A matrix containing three optimization metrics for all combinations
    of weights as provided by weights.df, or generated by 'step'.
    Automatically sorted by 'absolute'.
}
\description{
Calculate the optimum weights for each trait given a position, or group in
morphospace
}
\details{

}
